requirements.txt
matplotlib
numpy
Pillow
PyYAML
scipy
timm
torch
torchmetrics
torchinfo
torchvision
einops